requirements.txt
certifi==2020.6.20
chardet==3.0.4
gitdb==4.0.5
GitPython==3.1.8
idna==2.10
prompt-toolkit==3.0.7
requests==2.24.0
smmap==3.0.4
urllib3==1.25.10
wcwidth==0.2.5
